third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/keyframes-rules-expected.txt

Tests that source data is extracted correctly from stylesheets with @keyframes rules.

Content

Running: testInit

Running: testDumpStyles
[expanded] 
element.style { ()
    background-color: white;

[expanded] 
* { (keyframes-rules.html:9 -> keyframes-rules.html:9:4)
/-- overloaded --/     background-color: papayawhip;

[expanded] 
div { (user agent stylesheet)
    display: block;
